+++ builtin_commands.go
@@ -72,6 +72,7 @@ type builtinDocManCommand struct {
 	} `positional-args:"yes"`
 
 	builtinDocProgramNameOption
+	TrimDescriptions bool `long:"trim-descriptions" description:"Trim description whitespace in generated output"`
 
 	IncludeHidden bool `long:"include-hidden" description:"Include hidden options, groups and commands" description-i18n:"help.builtin.command.docs.include_hidden.desc"`
 	MarkHidden    bool `long:"mark-hidden" description:"Mark hidden entities in documentation output" description-i18n:"help.builtin.command.docs.mark_hidden.desc"`
@@ -81,6 +82,7 @@ func (c *builtinDocManCommand) Execute(_ []string) error {
 	opts := []DocOption{
 		WithBuiltinTemplate(DocTemplateManDefault),
 		WithProgramName(c.ProgramName),
+		WithTrimDescriptions(c.TrimDescriptions),
 		WithIncludeHidden(c.IncludeHidden),
 		WithMarkHidden(c.MarkHidden),
 	}
@@ -97,7 +99,8 @@ type builtinDocHTMLCommand struct {
 		Path string `positional-arg-name:"output" arg-name-i18n:"help.builtin.command.output.name" description:"Output file path" arg-description-i18n:"help.builtin.command.output.desc"`
 	} `positional-args:"yes"`
 	builtinDocProgramNameOption
-	TOC bool `long:"toc" description:"Include table of contents in output"`
+	TOC              bool `long:"toc" description:"Include table of contents in output"`
+	TrimDescriptions bool `long:"trim-descriptions" description:"Trim description whitespace in generated output"`
 
 	IncludeHidden bool `long:"include-hidden" description:"Include hidden options, groups and commands" description-i18n:"help.builtin.command.docs.include_hidden.desc"`
 	MarkHidden    bool `long:"mark-hidden" description:"Mark hidden entities in documentation output" description-i18n:"help.builtin.command.docs.mark_hidden.desc"`
@@ -113,6 +116,7 @@ func (c *builtinDocHTMLCommand) Execute(_ []string) error {
 		WithBuiltinTemplate(templateName),
 		WithProgramName(c.ProgramName),
 		WithTOC(c.TOC),
+		WithTrimDescriptions(c.TrimDescriptions),
 		WithIncludeHidden(c.IncludeHidden),
 		WithMarkHidden(c.MarkHidden),
 	}
@@ -129,7 +133,8 @@ type builtinDocMarkdownCommand struct {
 		Path string `positional-arg-name:"output" arg-name-i18n:"help.builtin.command.output.name" description:"Output file path" arg-description-i18n:"help.builtin.command.output.desc"`
 	} `positional-args:"yes"`
 	builtinDocProgramNameOption
-	TOC bool `long:"toc" description:"Include table of contents in output"`
+	TOC              bool `long:"toc" description:"Include table of contents in output"`
+	TrimDescriptions bool `long:"trim-descriptions" description:"Trim description whitespace in generated output"`
 
 	IncludeHidden bool `long:"include-hidden" description:"Include hidden options, groups and commands" description-i18n:"help.builtin.command.docs.include_hidden.desc"`
 	MarkHidden    bool `long:"mark-hidden" description:"Mark hidden entities in documentation output" description-i18n:"help.builtin.command.docs.mark_hidden.desc"`
@@ -148,6 +153,7 @@ func (c *builtinDocMarkdownCommand) Execute(_ []string) error {
 		WithBuiltinTemplate(templateName),
 		WithProgramName(c.ProgramName),
 		WithTOC(c.TOC),
+		WithTrimDescriptions(c.TrimDescriptions),
 		WithIncludeHidden(c.IncludeHidden),
 		WithMarkHidden(c.MarkHidden),
 	}